+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 2020-05-10
+- IR.CreateDevice - added
+- added jest tests (timers, variables, devices)
+
 ## 2020-01-15
 ### Added
 - IridiumMock.RemoveListener()